ui/gtk: Add scale option
Allow user to change initial scale of the virtual display. This is
helpful for users running QEMU on hi-dpi host desktop to achieve pixel
to pixel display -- e.g., if the scale factor of user's desktop is set
to 200%, then the user can set a 0.5 scale for the virtual display to
avoid magnification.

Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>