ui/gtk: Add keep-aspect-ratio option
When aspect ratio of host window and that of guest display are not
aligned, we can either zoom the guest content to fill the whole host
window or add padding to respect aspect ratio of the guest. Add an
option call "keep-aspect-ratio" to allow users to select their preferred
behavior in this case.

Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>